+++ src/cli/Microsoft.DotNet.UpgradeAssistant.Cli/Commands/UpgradeAssistantCommand.cs
@@ -14,14 +14,15 @@ namespace Microsoft.DotNet.UpgradeAssistant.Cli
         public UpgradeAssistantCommand(string name)
             : base(name)
         {
-            AddArgument(new Argument<FileInfo>("project") { Arity = ArgumentArity.ExactlyOne }.ExistingOnly());
+            AddArgument(new Argument<FileInfo>("project", LocalizedStrings.UpgradeAssistantCommandProject) { Arity = ArgumentArity.ExactlyOne }.ExistingOnly());
             AddOption(new Option<bool>(new[] { "--verbose", "-v" }, LocalizedStrings.VerboseCommand));
-            AddOption(new Option<IReadOnlyCollection<string>>(new[] { "--extension" }, "Specifies a .NET Upgrade Assistant extension package to include. This could be an ExtensionManifest.json file, a directory containing an ExtensionManifest.json file, or a zip archive containing an extension. This option can be specified multiple times."));
-            AddOption(new Option<IReadOnlyCollection<string>>(new[] { "--option" }, "Specifies a .NET Upgrade Assistant extension package to include. This could be an ExtensionManifest.json file, a directory containing an ExtensionManifest.json file, or a zip archive containing an extension. This option can be specified multiple times."));
-            AddOption(new Option<IReadOnlyCollection<string>>(new[] { "--entry-point", "-e" }, "Provides the entry-point project to start the upgrade process. This may include globbing patterns such as '*' for match."));
-            AddOption(new Option<UpgradeTarget>(new[] { "--target-tfm-support" }, "Select if you would like the Long Term Support (LTS), Current, or Preview TFM. See https://dotnet.microsoft.com/platform/support/policy/dotnet-core for details for what these mean."));
-            AddOption(new Option<bool>(new[] { "--ignore-unsupported-features" }, "Acknowledges that upgrade-assistant will not be able to completely upgrade a project. This indicates that the solution must be redesigned (e.g. consider Blazor to replace Web Forms)."));
-            AddOption(new Option<DirectoryInfo>(new[] { "--vs-path" }, "Path to a VS install directory to be used for %VSINSTALLDIR%. If not provided, the latest installed version will be used."));
+            AddOption(new Option<IReadOnlyCollection<string>>(new[] { "--extension" }, LocalizedStrings.UpgradeAssistantCommandExtension));
+            AddOption(new Option<IReadOnlyCollection<string>>(new[] { "--option" }, LocalizedStrings.UpgradeAssistantCommandOption));
+            AddOption(new Option<IReadOnlyCollection<string>>(new[] { "--entry-point", "-e" }, LocalizedStrings.UpgradeAssistantCommandEntrypoint));
+            AddOption(new Option<UpgradeTarget>(new[] { "--target-tfm-support" }, LocalizedStrings.UpgradeAssistantCommandTargetTfm));
+            AddOption(new Option<bool>(new[] { "--ignore-unsupported-features" }, LocalizedStrings.UpgradeAssistantCommandIgnoreUnsupported));
+            AddOption(new Option<DirectoryInfo>(new[] { "--vs-path" }, LocalizedStrings.UpgradeAssistantCommandVS));
+            AddOption(new Option<DirectoryInfo>(new[] { "--msbuild-path" }, LocalizedStrings.UpgradeAssistantCommandMsbuild));
         }
 
         protected class CommandOptions : IUpgradeAssistantOptions
@@ -47,6 +48,8 @@ namespace Microsoft.DotNet.UpgradeAssistant.Cli
             public IEnumerable<AdditionalOption> AdditionalOptions => Option.ParseOptions();
 
             public DirectoryInfo? VSPath { get; set; }
+
+            public DirectoryInfo? MSBuildPath { get; set; }
         }
     }
 }